refactor dispose workspace when using pvc
